outputs/sessions/48d6749d/files/requirements.txt
# This file is used by the Dockerfile to install dependencies.
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
pydantic==2.7.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4